Resolve #63 by updating function install
The format of the connection file which stores the parameters of the database connection was changed from a two column space seperated file to a two row comma seperated value. This change was made because the documentation was inconsistent with the use case and the csv format is easier to describe and is likely to be more familar to users.