Allow fetch to find temporary directory  (#132)
Due to a bug in the core retriever table names, including directories, were being lower cased:
https://github.com/weecology/retriever/issues/1101

Since `tempdir()` often includes uppercase letters this resulted in the core retriever trying to install
files into directories that didn't exist when using `fetch()`. The bug has been fixed in core, but since it is present in the last release this makes sure that the R wrapper works properly for retriever 2.1.0 by lower casing the output of `tempdir()`.

This could be removed in the future when the user based of 2.1.0 has dropped to sufficiently low levels.